fewer scripts w/ jasquat
