+++ spiffworkflow-backend/src/spiffworkflow_backend/scripts/get_url_for_task.py
@@ -28,7 +28,7 @@ class GetUrlForTask(Script):
         if human_task is None:
             raise ValueError(f"Could not find human task with guid '{task_guid}'")
 
-        public = kwargs.get("public", False)
+        public = args[1] if len(args) > 1 else kwargs.get("public", False)
         public_segment = "/public" if public is True else ""
         frontend_url = current_app.config["SPIFFWORKFLOW_BACKEND_URL_FOR_FRONTEND"]
         return f"{frontend_url}{public_segment}/tasks/{human_task.process_instance_id}/{human_task.task_guid}"
